Removed unused lines of code

+++ src/main/java/pl/shelter/shelter/animal/AnimalService.java
@@ -1,14 +1,7 @@
 package pl.shelter.shelter.animal;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.context.event.ApplicationReadyEvent;
-import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Service;
-import pl.shelter.shelter.meal.MealRepository;
-import pl.shelter.shelter.medicalCard.MedicalCardRepository;
-import pl.shelter.shelter.owner.OwnerRepository;
-
-import java.util.List;
 
 @Service
 public class AnimalService {
@@ -57,26 +50,17 @@ public class AnimalService {
         return animalRepository.save(animal);
     }
 
-    public void deleteAnimalById(Integer id) {
-
-
-        animalRepository.deleteById(id);
+    public void deleteAnimalById(Integer id) throws Exception {
+        if(animalRepository.existsById(id)) {
+            try{
+                animalRepository.deleteById(id);
+            }catch (Exception e){
+                throw new Exception ("Cannot delete animal by id", e);
+            }
+        }
     }
 
 
-    @EventListener(ApplicationReadyEvent.class)
-    public void fillDB() {
-
-/*        animalRepository.save(new Animal("Puma", "Dog", 8, 10, false, "1", "1"));
-        animalRepository.save(new Animal("Sonar", "Dog", 7, 10, false, "1", "1"));
-        animalRepository.save(new Animal("Paco", "Dog", 4, 15, true, "1", "1"));
-        animalRepository.save(new Animal("Hektor", "Dog", 17, 30, true, "1", "1"));
-        animalRepository.save(new Animal("Misiek", "Dog", 13, 16, true, "1", "1"));*/
-
-        System.out.println("females"+animalRepository.getNumberOfFemales());
-        System.out.println("males"+animalRepository.getNumberOfMales());
-        System.out.println("older"+animalRepository.getNumberOfOlderAnimals());
-    }
 
 
 }
